extract DoStateCheckpoint
