+++ execution/executor/src/components/mod.rs
@@ -12,6 +12,7 @@ pub mod in_memory_state_calculator_v2;
 pub mod chunk_result_verifier;
 pub mod do_get_execution_output;
 pub mod do_ledger_update;
+pub mod do_state_checkpoint;
 pub mod executed_chunk;
 pub mod partial_state_compute_result;
 pub mod transaction_chunk;
